requirements.txt
# Core API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration & Environment
pyyaml>=6.0
python-dotenv>=1.0.0

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
pypdf>=3.17.0

# Text Processing & Chunking
tiktoken>=0.5.1
langchain-text-splitters>=0.0.1

# Embeddings
openai>=1.3.0
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.35.0

# Vector Store
chromadb>=0.4.18

# Lexical Search
rank-bm25>=0.2.2

# LLM Clients
groq>=0.4.0

# RAG Evaluation
ragas>=0.1.0
datasets>=2.14.0

# Caching
diskcache>=5.6.3

# Utilities
httpx>=0.25.0
tenacity>=8.2.3
tqdm>=4.66.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Observability
python-json-logger>=2.0.7
prometheus-fastapi-instrumentator>=6.1.0
